Add support for null in histogram data source
There are occasionally some scenarios where the HotShot Query service has partial data
recorded for a block, but not all of the data.  In these scenarios some data can be missing for
a given block.  In particular, the missing data tends to be the payload of the block, which also
prevents the size of the block from being representable.

In order to allow for data in this partial state, the histogram data has been updated to allow for
the value `null` to be representable and accounted for within the histogram decoding process.