Update Summary Histogram decoding to allow for `null` values
Recently we have encountered a scenario where we run into an occasional error attempting to
retrieve the Explorer Summary from the Explorer API.  The error is caused by the confirmed block
header being stored before the payload of the block.  Due to this we can run into a scenario
where our resulting request retrieves a valid block header, but no block payload.  The server
side isn't expecting missing payload data, so this ends up returning an error.  To account for this eventuality, we need to allow for the histogram data to have `null` values
in the array of numbers.