Fix data differences between server and client
The server data allows for some `null` values in locations where they were not being
anticipated originally.  This has been adjusted with the introduction of a new function
that prefers the `null` value instead of the empty string value where applicable.

Additionally, some other members of classes have been adjusted to allow for a `null`
value to be passed through in them as well.