Remove command from `init.sh`
The init.sh script contains a floating Environment variable that was added
when modifying the `init.sh` script previously.  This Environment variable ends
up being evaluated as a command, which doesn't prevent anything from
working, but it does emit an error on launch to the logs.

This floating environment variable has been removed as it was never
intended to be in the `init.sh` script anyway.