Update `docker.yml` GitHub action to run on BuildJet
The current docker github action results in x86_64 binaries for the ARM builds
which is not what we would like / expect.  In an effort to address this, the runner
has been changed to target BuildJet instead, which supports ARM natively. As
a result, the need to setup QEMU has been removed, and the target, `runs-on`
field of the GitHub action has been removed.