Fix padding on explorer page sections
The padding on the explorer page's sections for "Latest Blocks",
"Latest Transactions", and "Most active rollups" is not what is
expected.  The titles are hugging the top of the card element too
closely.  This disparity isn't seen in the `Storybook` representation,
but it is on the NextJS build.

To address the inconsistency, additional styles have been added to
`typography.css` to remove the default margin values from header
elements.

To address the padding after removing header inconsistency, the
`card-padding` class has been added to some sub-elements, and
explicit padding has been added to the section titles.