Fix search bar showing over mobile menu
The search bar is currently drawn on top of the mobile menu when on the explorer page.
This is due to some z-index values being set for the search bar, and for the nav drawer
where the nav drawer's values aren't as large.

To address this the nav drawer simply needs some larger z-index value.  10 is chosen here
in order to try and avoid some potential future z-index elements that could be introduced
by the search bar.  It is larger than it needs to be, but it also gives room for elements
with additional z-index values.