Fix padding on Transaction Pages
The padding on the Transactions pages is inconsistent.  After making the adjustment
to the typography.css file, the heading elements all have no margin.  Additionally, the horizontal spacing of the `TableLabeledValue` is very tight / non-existant on
the edges of the section cards.

This adjustment adds card padding to the `TableLabeledValue` components that are
present on the transaction pages.