Create mock tests for HotShot Query Service
The class created to be the abstraction over the HotShot Query Service
is not currently effectively tested.  This is primarily due to needing a live
service to be able to serve responses, which we cannot guarantee.

To get around this issue we provide a mock implementation of the
`fetch` function that can intercept these calls and return a specific
response for them.