Remove single line function
Based on discussion with @Sneh1999:
https://github.com/EspressoSystems/espresso-block-explorer/pull/154#discussion_r1801561403

The function `createFetcherSuccessResponse` is a single line, and does
not seem to add significantly easier use to the function it is wrapping. As
a result, it doesn't really need to exist, and can be replaced with the the
call to the function it is wrapping anyway.