Add Inscription service definitions
The Inscription service represents a toy app demo for interfacing with
Espresso's block chain.  The point of the Inscriptions API is to provide a
simple feature set for interacting with, and demonstrating the Espresso
Block Chain.

This is done by allowing the user to connect his / her wallet to the page
and sign an affirmation of commitment to the "infinite garden".