Modify friendly date time format
The friendly date time format appears on a wall of inscriptions.  When multiple dates
and hours are next to each other, they look to line up in the majority of cases, due
to the monospace font.  But there are times where it does not line up.  This is due to
single digit versus double digit representation of times.

Examples:

- Nov 6, 2024, 12:34 PM MST
-  Nov 6, 2024, 1:56 PM MST
- Nov 10, 2024, 1:56 PM MST
- Nov 10, 2024, 12:34 PM MST

The easiest way to address this is to have the hours, and the day portions of the time
stamp have two digits, instead of being auto formatted.

This modifies the friendly date time format such that the day portion, and the hour
portion are two digits.