Fix "data lines" vertical spacing
From the Espresso Block Explorer Design QA document provided by Martin
Ahe.

Martin states that the data lines in the block and overview areas are supposed
to have 8px of vertical spacing.

In this case he actually means to say that the spacing between the elements
in these sections should have a vertical separation between the items of 8px.

This adjusts these values to have the correct spacing