Fix section title colors and alignment
From the Espresso Block Explorer Design QA document provided by Martin
Ahe

Martin asserts that the "Latest Blocks" and "Latest Transactions" section do
not match the correct color of the specification, and that they do not have the
correct alignment within the container.

This change modifies the color of these titles to be `slate-950` as intended,
and to center their vertical alignment within the container.