Fix copy button spacing
From the document Espresso Block Explorer Design QA provided by Martin
Ahe

In the document martin asserts that the design for the copy symbols indicates
that they should always have a 4px gap.

These changes adjust this gap so that the button can be positioned more
correctly when inline with other text.