Adjust card padding on multiple elements
The `Card` element was was lacking its original padding that was the purpose
of the `CardNoPadding` element.  This caused some visual inconsistencies in
several places.

These inconsistencies have been fixed by adding the `card--padding` class
to the `Card` component.  As a result of this change, several consumers of
the `Card` component have been adjusted accordingly.