Update error display when lacking a better component
As a catch-all fail safe, when an error occurs that we do not have a dedicated
error display for we display some text that is considered tot be not user-
friendly.  It has been requested that this error message be modified to
something that is more appropriate for the general public to consume rather
than the message it contains now.

The error display has been modified to indicate that there was a general
failure due to a technical issue, and that our team is currently working on it.

I fear that this may be misleading as this is an error that is going to be
displayed to the user, and it may imply that we are currently aware of the
particular issue that the user is experiencing, when we may not be.  We
don't get notified, nor are we capable of being notified currently, when a
user experiences an error message that we didn't account for.  The only
ones we might know about are issues that stem from the server itself,
which I would not classify as an "Explorer issue".

In any case, this error message has been updated to reflect the sentiments
and request as appropriate.