Fix search input tests data source
With the change of the Adapters to no longer provide an initial implementation
on load, the previously written storybook tests and UI no longer return any
results as they have no implementation specified.

This adds a component to provide the fake data implementation for the
tests so that they can access the information needed, and return actual
results.