Add search input activation tests
The search input currently does not have any tests that check to see whether
an activation of submission to the input, using the enter key, will actually
attempt to perform a navigation or not.

This refactors the Search Controller to take a passed `Location` object to
operate on instead of attempting to grab the window's.  This change is
backwards compatible as the parameter is optional, and if not specified
it will fall back to the relevant target window location.

A MockLocation class has been added that conforms to Location, so that
Operations on it will behave as expected.