Add company website to node identity
The Node Identity information retrieved from the node validator service can
contain a public facing website that should be listed along with the other
node identity information in the table of nodes.

This change updates the processing of the node identity to populate the
company website, and then utilize it to populate the node information in
the table.