Update default sort on nodes table
Some feedback given from @Ancient123 is that the nodes table is sorted by
the public key by default, which doesn't really make a lot of sense, since the
data will end up looking like it's in a consistent, but random, order.  Instead he
suggests that we should sort by Vote Participation instead, as this is a more
useful statistic for us at the moment.

This change modifies the sort direction of the Nodes Summary Table to
sort by vote participation by default instead.