Update the InscriptionsPage smoke test to account for having content
The initial smoke tests didn't ensure that we had some valid inscription
entries on the page.

In attempting to resolve this issue it was noticed tha we are registering
some codec responses more than once for the Inscriptions Requests.
This is likely due to the way web workers are handled in the testing
harness.  The error has been removed in favor of doing nothing silently.

This modifies the InscriptionsPage smoke test to ensure that some
relevant content is being rendered, by checking that inscription
elements are present on the page.