Add further checks to the NodesPage smoke test
The NodesPage smoke test doesn't currently test any of the functionality and
rendering.  Instead it just checks that the initial render works, and nothing
else matters.  This test should check to ensure that the various card sections
load and display information as expected.

This change adds extra checks to the NodePage smoke test to ensure that
the data is being processed, loaded, and that the components are being
displayed without issue.