Add a loading placeholder for histograms
When the histograms are in a loading state they currently do not display
anything.  This makes them take up no space, and causes a layout shift when
the page is finished loading.  This can lead to a bad user experience when the
page finishes retrieving the data it needs.

What the page needs is for the histograms to have a consistent amount of
space being given to it when it is loading, or at least as consistent as we
can make it.  Since the histograms themselves adhere to an aspect ratio
we should be able to compute the size of the space with ease using this
ratio, and some css tricks.

This change adds loading displays for the histograms, and adds a test
case for the loading state of the histogram.  As a result, the card portion
of the histogram has been moved from being supplied at the Nodes
Page to the histogram itself.  It may be modified in the future to allow for
the class name to be passed in from the NodesPage.