Fix Dockerfile build process
With the addition of the Inscriptions demo to the components library, the build
process for the block explorer Docker image broke.  Additionally, when adding
the ability to test search input interactions, we introduced a naked `window`
reference that also breaks the build.

These changes fix the Docker build process, and the NextJS build process for
the block explorer.