Replace Link with separate Link types
Many React frameworks provide custom Link components that allow
for preloading capabilities when navigating around on the webpage.
These components can provide an optimization when navigating
around as the page itself doesn't need to be replaced, or reloaded.
This in turn leads to faster page navigation, seamless loading, and
a better overall user experience.

This isn't a universal rule, however, and some things need to be able
to operate without this Link by using just what is available via html
itself.  Things such as storybook would require us to be able to
navigate with just raw anchor tags for the stories presented.

As a result, we find ourselves needing to potentially replace the
link component with a custom drop-in replacement, while also
defaulting to the base anchor component.  Additionally, we only
want to to this sort of navigation when we are navigating to another
part of the known website.  So we also need to separate links
that stay on the website domain  (`InternalLink`s) versus links that
leave the website (`EgressLink`s).

This change adds these links, and replaces the previous link
reference with the more appropriate link.

This change also adds a Context that allows the user to override the
Link component to use for InternalLinks.

This change also replaces the block-explorer's default InternalLink
with NextJS's Link component.