Modify WebWorkers to avoid creating duplicates
When implementing the changes to the Link in order to separate
`InternalLink`s from `ExternalLink`s, it was noticed that navigating
around in the `NextJS` project that navigating back and forth
between the `Explorer` and `Nodes` pages would spawn a new
fresh `WebWorker` every time, instead of reusing the existing one.

This is inherently wasteful, as we would end up spawning
WebWorkers that will live in their own thread, but never be utilized
again.  In essence, they will just take up memory space, and not
be useful going forward.