Modify provided web worker instances to be singleton
In experimentation with the project when allowing for custom Link
elements to utilize NextJS's `Link` component, it was discovered that
navigating back and forth between pages would result in numerous
Web Workers being created to perform the same task, without any clean
up. Since these Web Workers hold state, and don't get removed, we
could end up with several instances all of which are holding memory
and a thread open for processing.

This is undesirable, as it is a memory and resource leak.

This happens due to the `Link` element of `NextJS` causes the overall
page to be modified instead of reloading in order to facilitate a better
user experience.  As a result, the solution needs to be aware of this
potential, and needs to be adjusted to consider this persistent
environment.

As a result, the creation of the Web Worker services have been modified
to return singleton instances of their implementations, so that they won't
be recreated unnecessarily.