Fix multiple consumers of service event streams
With the change to prioritize singleton instances for the Web Worker
services, it was noticed that upon navigating back and forth between
the Nodes Page, that we would start losing some events.  It was
determined that the cause was due to bridging the stream into multiple
consumers.  We do this in order to ensure that we process the events
and send them downstream.  However, we encounter an issue in that
we never stop doing this when the component unmounts.  As a result
these unmounted consumers will continue indefinitely intercepting
events from mounted consumers.

To fix this issue, we implement a signal that can inform the consumer
to stop processing on unmount.  This has some implications as we will
need to consider two separate Promise values at once.  We utilize
Promise.race for this, but we also need to be aware of the potential
to lose events as a result.  Additionally, when we do close the stream,
we need to drain the remaining events until we receive the close event
we expect.

This change implements these considerations and logic.