Fix WebSocket command processing in invalid states
When React is in dev mode, it utilizes strict mode in order to try and
ensure that user's hook handling are implemented correctly.  This caught
an issue with the WebSocket command processing.  We expect these
events only to come through when the Web Socket is in a "valid" state.
With the strict behavior the WebSocket commands "Open" "Close"
"Open" are sent very close to each other.

Initial attempts were to try and work around this limitation rather than
address the handling of these events in cases where they are not
expected.  We were receiving a "Close" when we weren't technically
finished being "Opened", and likewise another "Open" when we weren't
fully "Opened" or "Closed".

The correct solution is to account for these states more accurately.  We
can still ensure that we are in the correct state before we operate, but we
should still handle and provide a valid response when we are in these
unexpected states.

So  we implement better condition handling for these edge cases that
ultimately help to address the underlying issue.