Modify WebSocket handlers to behave better with async calls
When in dev mode, the WebSocket handling code runs into issues when
React strict mode activates and rapidly triggers a mount, unmount, and
mount again.  We ultimately want to be able to handle this behavior, as it
being directly triggered like this indicates what could happen in a real
world scenario, even if unlikely.

To address and better handle this behavior, the handling methods in
the WebSocket have been made to await for the individual tasks to
complete before attempting further tasks.  This will ensure that our
websocket behaves as expected when rapidly given commands that
require the previous commands to complete.

This change allows the Node page to function correctly in both dev
mode, and release mode in NextJS.