Fix flakiness of Async Iterable Resolver Component test
The test fo the Async Iterable Resolver Component has been
occasionally failing due to the number of events that are being triggered.
In reviewing the test, it seems that some of this may be due, in part, to
the rerender trigger that has been happening.  Additionally, the waitFor
conditions may not have been sufficient in order to test the behavior
as precisely as needed.

This change addresses these concerns by removing the rerender, and
by adding my `waitFor` conditions, in order to ensure that we are
hitting the correct states before verifying the recorded states.