Remove require call in storybook main file
There is an error that occurs when attempting to run storybook with some newer
versions of Node.js.  This error is caused by a call to `require.resolve` which ends
up in the error being thrown.  This is likely related to using es modules rather than
commonjs.

This change fixes this by removing the `getAbsolutePath` function call that
resolves the path to the project / workspace root.  This no longer seems to be
necessary, and removing it resolves the issue.