Fix "act" warning when running testes
When running tests on the Block Explorer a bunch of warnings concerning "act"
are displayed.  This is due to a difference in `userEvent` provided by storybook
versus the testing-library package, and for not awaiting the results of these
user event interactions.

This change replaces the userEvent usage with the import from `testing-library`, and
ensures that the results of these interactions are awaited in async code.