Add more info to block time to add context
We have had multiple questions as to why our block time does not match our advertised
block time.  We've received this question enough times that we felt it was prudent to
add an explanation to the block time to help users understand why the block time may
differ from our advertised optimal block time.

This commit adds a few new components to facilitate this:
- Add MoreInfoElement
- Add HoverDialog
- Add PrefixMoreInfoElement

These components are utilized to provide an "(i)" icon element that can be hovered to
provide additional context to the user.