Update Node Validator Adapters to support data update
With the modifications to the Block time histogram, requiring the data for the blockSize
in addition the blockTime, the histograms on the Nodes page broke.  This is due to the
histogram data being provided explicitly in the `publishHistogramUpdates` function.

To fix this issue the requirements for the blockTime data need to be amended to also
include the blockSize data.  This will more accurately reflect the data requirements
for the histogram data.

This commit makes that adjustment and populates the blockSize data for blockTime
as well.