Adjust the histogram styles on the Nodes page
With the changes made to the Block Time Histogram, the styles also need to be carried
over. The style changes needed for the Nodes Page are the same ones that allow for the
header value to be set at the top.

This commit adds the same styles to the Nodes Page Histogram so that the histograms
themselves are pushed to the bottom of the available space.