Clean up logic in `BlockTimeHistogram`
The logic in `BlockTimeHistogram` is a bit dense when it comes to computing the
averages for block times with empty and non-empty blocks.  The existing code is
also contains a few type annotations that are necessary to appease typescript.  This
results in code that's a bit too dense to read effectively, as there's too much happening
in the local context.

This has been cleaned up by extracting a couple of key functions to help improve the
code's readability, and boil down the differences to a single passed anonymous function.
All other matters are handled by the extracted functions, which are able to much more
clearly express the intent of the code.

Add `nullableMean` getter to `DataStatistics`

The `DataStatistics` `mean` value is guaranteed to be non-null, as it's computed from
the sum, which starts as a `number` already.  This makes the user need to inspect the
`count` value to determine whether there actually is, or should be, any mean value or
not.

This change introduces the `nullableMean` getter as a convenience to the user, which
will return `null` if there are no values in the dataset.