Update block time alignment for resampled block times
Since the Builder is often waiting for transactions before it submits a block, we end up
with some block times that are on the falling edge of the available block build timeout.
This doesn't really capture how quickly we can produce blocks.  At the recommendation
of @jbearer we should sample the time of the next block, as that will be within the period
of eager block production to expedite finalization.

This change does just that by adjusting the alignment of the block time for the resampled
blocks.  It offsets each block `N` to correspond with the block time of block `N + 1`
instead.