Add Explorer wallet connect components
In order to be able to allow the user to delegate to validators, we need
wallet interaction in the Block Explorer.  We can start by adding some
basic components for the user to login, and see which account he/she
is currently connected with.

This commit adds some basic components for connecting the user's
wallet.

This commit also refines the type definitions of the information provided
by RainbowKit.  This is done by deriving the types using Typescript's
type extraction primitives.