Add RainbowKitMountedGuard component
In order to be able to utilize RainbowKit, it must be initialized and
mounted first.  If it is not, it can cause errors to be thrown when
attempting to access properties that would otherwise be assumed to
exist.

This component is convenient in that it adds a simple conditional
clause that will only render the specified children if the RainbowKit
scripts have been loaded, and successfully mounted.