Remove stand-alone inscriptions project
The inscriptions campaign, known as the Infinite Garden was a campaign
that was short lived to drive traffic to the Espresso Mainnet.  The
campaign Occurred in November of 2024, and no longer needs to be
maintained.  The Inscriptions themselves are able to be decoded and
displayed on the Block Explorer accurately.  There is no longer any need
to maintain the separate components, pages, and packages for them.

As a result they have been moved, and their underlying data that is still
referenced has been relocated within the project.