Add alignment configuration to DataTable
When displaying table information, it is usually expected that the
informational values are left (start) aligned, and the numeric values
are right (end) aligned.  This helps to make the information being
presented to be better understood visually at a glance.

In order to be able to facilitate this, we need the ability to specify the
alignment of the cells themselves.

This commit adds the option to specify the alignment for the DataTable
accordingly.