Add configuration definitions and contexts
In order to extends the Nodes Page to provide Web3 functionality for
delegating stake, we need to be able to derive the relevant information
about the contracts and geth instances we are working with.  We need
to be able to make this configurable as we will need to support different
deployed environments.  Additionally, it would be beneficial in the
development settings to be able to override which URLs we are
interacting / querying against for the HotShot Query Service, and the
Node Validator services.

This commit supplies three different environment / configuration
definitions and Contexts to make this information available to the
entire React tree.