Add Keccak hash implementation
The keccak algorithm is useful to have access to within the project.  We
already have an indirect dependency on the keccak repository, however
that library prioritizes the use of native FFI function calls with a fallback
to a Node Based implementation.  The Node Based implementation
makes use of the `Buffer` type which is not available to browser
environments.  This makes the typescript types not happy when
attempting to interact with the module in any form.

As a result the Keccak library has been adapted to be implemented
natively using the appropriate typescript definitions, with its
implementation coming directly from the keccak javascript fallback.