Modify ArrayBuffer Codec to add additional support
The ArrayBuffer Codec has served us well so far, but it needs to be
extended / expanded to be able to support additional types.  There
are types that look and seem to have a class that is called an
`ArrayBuffer`,  but seems to be different than the globally available
`ArrayBuffer`.  As a result, the instanceof check fails in these cases.

This requirement has been loosened in order to support these different
ArrayBuffer classes.