Add EIP 55 Codec
The EIP 55 format helps to try and ensure valid address data entry by
modifying the case of hex characters depending on a hash of the
address itself.  This allows for a checksum to be embedded into the
representation of a wallet address without modifying the data that
the underlying address represents.

This format is widely accepted and utilized.  For consistency we
should utilize its representation in our explorer as well.  This
coded provides the ability to do just that.