Add Stake Table definition
In order to be able to receive a Stake Table from the Node Validator
service / Hot Shot Query Service, we need to be able to decode the
Stake Table.  As a result, we need a local definition of the Stake
Table that the UI can represent.

This commit adds the definition of the Stake Table for the application
to consume and display accordingly.