Add SetURL commend to WebSocketCommands
In order to support a development environment overridable
WebSocket URL, we need to be able to modify the URL that the
web worker will connect to.  Normally this is handled / controlled by
the config file.  However, in order to be able to support multiple
environments, we need to be able to change this based on
user supplied configurations within the Storybook setup.