Add RewardState to HotShotQueryService API
In order to be able to show the user his/her current rewards we need to
be able to retrieve these rewards.  These rewards are currently only
available on the HotShot Query Service, as they are not directly
tracked within the contract itself.

The commit adds the RewardState sub-api for the HotShotQuery
Service, in order to be able to query this information.