Update WebSocket Handling to support rapid reconnect protection
Due to the nature of how surf-disco APIs handle unrecognized client
requests, it would be beneficial to have an exponential back-off to
WebSocket reconnect requests.

When the WebSocket sends a command that the server does not
recognize, the server will just close the client connection.  With
automatic reconnect support, this can end up in a rapid-fire pseudo
hot-loop of spamming the WebSocket with reconnect requests.

This commit adds protections from this sort of behavior through
indirect detection of such a failure.