Add ability to overwrite the URLs for the web worker services
In order to be able to test different endpoints that can be configured
dynamically (for storybook) we need to be able to supply a URL for
the Web Workers to interact with dynamically.

This commit adds support for these changes, as well as adds
synchronized serialization processing of requests that are coming in.

This should help to ensure that the requests are processed sequentially
in order to avoid race conditions.