Add TextEditing based input
The text editing capabilities of Javascript leave a lot to be desired. In
order to be able to improve upon this interaction model it is beneficial
to take a unified state derived approach.  It is also beneficial in many
cases to provide the user with non-destructive transformation
capabilities.  However, this can be difficult to achieve on their own, and
the user may still want to overwrite or restrict what a user is, and isn't
able to enter.

This commit adds some initial support for a Text Editing approach that
is similar to that provided by Flutter.  It aims to make these edits
almost entirely state generated.