Modify the Nodes Page to support Stake Delegation
We would like for the user to be able to see current information about the
Validators that can inform him/her about the delegation opportunities
that are available.  Additionally, we would like for the user to have an
interface to be able to easily and effectively choose which nodes to
delegate their ESP tokens against.

This change is fairly large, but it unifies all of the modifications to the
Nodes Page in order to be able to support the desired delegation
interaction we are looking to be able to support.