Upgrade Storybook, NextJS, React, and others
The upgrade to all dependencies required a lot of code adjustments. This
was primarily due to Storybook and some of it's package / semantic
changes.  As a result, this update touches a quite large number of files
within the repo.

These changes were applied with the migration tool, and then manually
adjusted to make things work.

The modification to the NextJS project modified the components to be
ServerComponents where possible, and then to call to Client
Components where necessary.  This required a clean separation between
the server-side dependencies and the client side dependencies. The
espresso-block-explorer-components library can only be consumed on
the client side, so care needs to be taken in this separation.