Fix linting configuration
The eslint package has completely changed how you manage its
configuration and settings.  This is causing the lint test to fail every time
it is run.  The package and configuration need to be brought up-to-date.

This commit has utilized the eslint migration utility to convert the
previous configuration to the new one.  Additionally, it removes the
previous file as it is no longer needed.  Some additional rules have
been added to disable lint errors on deprecated React features.