Update build / launch process for the block explorer
NextJS prevents us from being able to deliver dynamic content when the
user accesses the server at run time.  You are able to reference
environment variables in NextJS, and they can be retrieved in
ServerComponents, whose results are then passed to the
ClientComponents.  However, when we `build` the next bundle, the
environment variables get replaced with their baked values.  This
means that in order for us to dynamically configure the environment
we will need to perform the build step for `next` right before we
`start` the server.

This commit makes the necessary image adjustments to achieve these changes.