Fix vitest configuration for block-explorer package
vitest seems to be unable to resolve the typescript imports when running
tests.  However, it is able to resolve them when building the project.
Based on a search for the error, a stack overflow post indicated that the
reason is that the tsconfig has defined the alias, but vitest still needs to
be informed about it.  The solution is to use a plugin for tsconfig paths
for vitest.