Add Label component and replace <label> usage
The Web Inspector tools were complaining about the use of the HTML5
`label` element when it doesn't correspond to a form input.  The `label`
element isn't meant to be utilized for a typographical label by itself, or at
the very least not without pointing at an input element.  To resolve this
issue it is better to create a custom `Label` component that has the
styling desired, rather than to keep miss-using the existing `label`
element.

This commit Adds the `Label` component and replaces instances of
the `label` usage with it instead.
Add missed Label change