Clean up some Stories
There are some stories that are not displaying exactly as expected.
Specifically when it comes to the utilization of the horizontal divider and
their Storbyook component layout sections.

This commit addresses those issues by replacing use of the `hr` element
with the `Divider` element, and by correcting the errant Storybook path.