Modify world dots to scale based on population
@jbearer had requested a feature to differentiate the display of the world
map dots based on the population of the node.  Since this is a direct
comparison of scales, it makes the most sense to scale the dots as
a result.

However, this can lead to complication, as everything is grid-aligned, it
has the potential for adjacent nodes not to be able to be inspected.  To
account for this, it makes sense to take advantage of browser / svg
rendering behavior to layout these element in the order that they are
drawn.  If they appear first in the DOM, that means that they'll be
behind elements that come after.  Due to this, it becomes beneficial to
sort the dots based on their population so that the large nodes appear
in the background where they should still be accessible if they are
overlapped by a smaller dot.