Add Nitro Integration display helpers to block explorer
Due to working with the nitro team on their integration I have been exposed to the
encoding format of the data posted by their integration.  As a result, we have the
ability to easily decode their format and present it on the explorer for reference.

This can be beneficial in displays the contents of a known block format and what it
contains.  This will be great for debugging, troubleshooting, and just general
information.