Rename strings for Nitro based integrations
Based on feedback I received from @Sneh1999 and @zacshowa in a slack discussion,
some fields could be renamed to add clarity to the information being presented.  The
fields specifically mentioned were to change `Block Number` to `L1 Block Number`,
and to change `Index` to `Rollup Block Number`.

https://espressosys.slack.com/archives/C08R249J2HJ/p1755535024835449?thread_ts=1755322153.037289&cid=C08R249J2HJ
https://espressosys.slack.com/archives/C08R249J2HJ/p1755535150789319?thread_ts=1755322153.037289&cid=C08R249J2HJ

This change replaces the strings used for the labels of these fields.