Adds rollup identification and images for t3rn
The t3rn chain has appeared on mainnet.  In order to display this appropriately, their
relevant chain information has been added to the curated list of rollups.

Fix LogX for dark mode

The LogX logo is black, and in dark mode it becomes invisible.  In order to address this
The LogX background color has been added to their logo to reflect how their logo
looks where they display it.  This background makes it possible for the same logo to
be utilized for both dark and light pages.