Split Delegation UI to its own page
Based on feedback and discussions from Lucca

Adding the delegation UI to the Nodes page seems to add a lot of
visual noise and clutter.  It would be better for the Delegation UI
to have its own page so it can be optimized for that specific
use-case.

Name adjustments have been made for clarity as well. Preferring
terms like "Stake" over "Delegate".