Update styling for `HexDump`
The entire hex dump area is selectable and copyable without being
able to easily distinguish between the byte offset section, and the
byte contents themselves.  It could help to be able to select these
fields independently, as it's likely unnecessary to grab the byte
offset when attempting to copy or read data.

This change splits the byte offsets from the byte contents
themselves, and adds a separate coloring track for the byte
offsets.